+++ src/_util/noop.ts
@@ -1,2 +1,4 @@
 // eslint-disable-next-line @typescript-eslint/no-empty-function
-export default () => {};
+const noop = () => {};
+
+export default noop;