+++ src/button/Button.tsx
@@ -7,12 +7,14 @@ import Loading from '../loading';
 import { TdButtonProps } from './type';
 import { buttonDefaultProps } from './defaultProps';
 import parseTNode from '../_util/parseTNode';
+import useDefaultProps from '../hooks/useDefaultProps';
 
 export interface ButtonProps
   extends TdButtonProps,
     Omit<React.AllHTMLAttributes<HTMLElement>, 'content' | 'shape' | 'size' | 'type'> {}
 
-const Button = forwardRef((props: ButtonProps, ref: React.RefObject<HTMLElement>) => {
+const Button = forwardRef((originProps: ButtonProps, ref: React.RefObject<HTMLElement>) => {
+  const props = useDefaultProps(originProps, buttonDefaultProps);
   const {
     type,
     theme,
@@ -94,6 +96,5 @@ const Button = forwardRef((props: ButtonProps, ref: React.RefObject<HTMLElement>
 });
 
 Button.displayName = 'Button';
-Button.defaultProps = buttonDefaultProps;
 
 export default Button;