+++ src/date-picker/DateRangePicker.tsx
@@ -18,11 +18,14 @@ import {
 import { subtractMonth, addMonth, extractTimeObj } from '../_common/js/date-picker/utils';
 import { dateRangePickerDefaultProps } from './defaultProps';
 import log from '../_common/js/log';
+import useDefaultProps from '../hooks/useDefaultProps';
 
 export interface DateRangePickerProps extends TdDateRangePickerProps, StyledProps {}
 
-const DateRangePicker = forwardRef<HTMLDivElement, DateRangePickerProps>((props, ref) => {
+const DateRangePicker = forwardRef<HTMLDivElement, DateRangePickerProps>((originalProps, ref) => {
   const { classPrefix, datePicker: globalDatePickerConfig } = useConfig();
+
+  const props = useDefaultProps<DateRangePickerProps>(originalProps, dateRangePickerDefaultProps);
 
   const {
     className,
@@ -254,8 +257,9 @@ const DateRangePicker = forwardRef<HTMLDivElement, DateRangePickerProps>((props,
   }
 
   // 确定
-  function onConfirmClick() {
+  function onConfirmClick({ e }) {
     const nextValue = [...inputValue];
+    props?.onConfirm?.({ e, date: nextValue, partial: activeIndex ? 'end' : 'start' });
 
     const notValidIndex = nextValue.findIndex((v) => !v || !isValidDate(v, format));
 
@@ -275,12 +279,12 @@ const DateRangePicker = forwardRef<HTMLDivElement, DateRangePickerProps>((props,
     }
 
     // 首次点击不关闭、确保两端都有有效值并且无时间选择器时点击后自动关闭
-    if (!isFirstValueSelected) {
+    if (!isFirstValueSelected || nextValue.length === 1) {
       let nextIndex = notValidIndex;
       if (nextIndex === -1) nextIndex = activeIndex ? 0 : 1;
       setActiveIndex(nextIndex);
       setIsFirstValueSelected(true);
-    } else {
+    } else if (nextValue.length === 2) {
       setPopupVisible(false);
     }
   }
@@ -377,6 +381,5 @@ const DateRangePicker = forwardRef<HTMLDivElement, DateRangePickerProps>((props,
 });
 
 DateRangePicker.displayName = 'DateRangePicker';
-DateRangePicker.defaultProps = dateRangePickerDefaultProps;
 
 export default DateRangePicker;